implement proposed folder structure
