+++ src/boiler/pages/Content.jsx
@@ -1,34 +1,34 @@
 import React from 'react';
-import {ReactComponent as ArrowRight} from '../svg/arrow-right.svg';
+import { ReactComponent as ArrowRight } from '../svg/arrow-right.svg';
 
 export const colors = [
-    "aqua",
-    "blue",
-    "navy",
-    "teal",
-    "green",
-    "olive",
-    "lime",
-    "yellow",
-    "orange",
-    "red",
-    "fuchsia",
-    "purple",
-    "maroon",
-    "white",
-    "silver",
-    "gray",
-    "black"
+    'aqua',
+    'blue',
+    'navy',
+    'teal',
+    'green',
+    'olive',
+    'lime',
+    'yellow',
+    'orange',
+    'red',
+    'fuchsia',
+    'purple',
+    'maroon',
+    'white',
+    'silver',
+    'gray',
+    'black',
 ];
 export const brandColors = [
-    "primary",
-    "secondary",
-    "success",
-    "danger",
-    "warning",
-    "info",
-    "light",
-    "dark"
+    'primary',
+    'secondary',
+    'success',
+    'danger',
+    'warning',
+    'info',
+    'light',
+    'dark',
 ];
 
 const ContentPage = () => {
@@ -47,7 +47,10 @@ const ContentPage = () => {
                 <div className="swatch-list mb-3">
                     <div className="row">
                         {colors.map(color => (
-                            <div className="col-xs-6 col-md-2 txt-center">
+                            <div
+                                key={Math.random() + '_color'}
+                                className="col-xs-6 col-md-2 txt-center"
+                            >
                                 <div className={`swatch bg-${color}`}></div>
                                 <p>
                                     <code>${color}</code>
@@ -61,7 +64,10 @@ const ContentPage = () => {
                 <div className="swatch-list">
                     <div className="row">
                         {brandColors.map(color => (
-                            <div className="col-xs-6 col-md-2 txt-center">
+                            <div
+                                key={Math.random() + '_color'}
+                                className="col-xs-6 col-md-2 txt-center"
+                            >
                                 <div className={`swatch bg-${color}`}></div>
                                 <p>
                                     <code>${color}</code>
@@ -314,7 +320,6 @@ const ContentPage = () => {
                         </tr>
                     </tbody>
                 </table>
-
             </section>
         </>
     );
